Revert "feat(panels): dockable dashboard (B2)"
This reverts commit c04cd3c287df84ef311b5a8acde33541d9bf8164.
