+++ src/App.tsx
@@ -1,33 +1,14 @@
-import { useState, useEffect, useMemo, useCallback } from 'react';
+import { useState, useEffect, useMemo } from 'react';
 import { type ChartDataset } from 'chart.js';
-import type { DockviewApi } from 'dockview-react';
+import DateRangeSelector from './components/DateRangeSelector';
 import Footer from './components/Footer';
 import Header from './components/Header';
-import DockShell, {
-  PANEL_DEFS,
-  PANEL_IDS,
-  type PanelId,
-} from './dock/DockShell';
-import {
-  DockLayoutProvider,
-  type DockLayoutContextValue,
-} from './dock/DockLayoutContext';
-import {
-  DashboardProvider,
-  type DashboardContextValue,
-} from './context/DashboardContext';
-import ChartPanel from './dock/panels/ChartPanel';
-import DateRangePanel from './dock/panels/DateRangePanel';
-import LineSelectorPanel from './dock/panels/LineSelectorPanel';
-import MapPanel from './dock/panels/MapPanel';
-import PanelChrome from './dock/panels/PanelChrome';
-import SummaryPanel from './dock/panels/SummaryPanel';
-import useIsDesktop from './dock/panels/useIsDesktop';
+import LineSelector from './components/LineSelector';
+import OutputArea from './components/OutputArea';
 import useUserDashboardInput, {
   type UserDashboardInputState,
 } from './hooks/useUserDashboardInput';
 import { getLineColor, getLineNames } from './utils/lines';
-import { clearLayout } from './utils/layoutStorage';
 import type { CustomChartData } from './@types/chart.types';
 import type {
   ConsolidatedRidership,
@@ -35,43 +16,9 @@ import type {
 } from './@types/metrics.types';
 import ridershipRecords from './data/ridership.json';
 
-const allPanelsVisible = Object.fromEntries(
-  PANEL_IDS.map((id) => [id, true]),
-) as Record<PanelId, boolean>;
-
-/** Rebuild the default layout; mirrors DockShell's initial build. */
-const buildDefaultLayout = (api: DockviewApi): void => {
-  for (const id of PANEL_IDS) {
-    const def = PANEL_DEFS[id];
-    api.addPanel({
-      id,
-      component: def.component,
-      title: def.title,
-      ...(def.position ? { position: def.position } : {}),
-    });
-  }
-
-  for (const id of PANEL_IDS) {
-    const { defaultHeight, defaultWidthRatio } = PANEL_DEFS[id];
-    const panel = api.getPanel(id);
-    if (!panel) continue;
-
-    if (defaultHeight !== undefined && api.height > 0) {
-      panel.api.setSize({ height: defaultHeight });
-    }
-    if (defaultWidthRatio !== undefined && api.width > 0) {
-      panel.api.setSize({ width: Math.round(api.width * defaultWidthRatio) });
-    }
-  }
-};
-
 function App() {
   const [isLineSelectorExpanded, setIsLineSelectorExpanded] =
     useState<boolean>(false);
-  const [dockApi, setDockApi] = useState<DockviewApi | null>(null);
-  const [panelVisibility, setPanelVisibility] =
-    useState<Record<PanelId, boolean>>(allPanelsVisible);
-  const isDesktop = useIsDesktop();
 
   const userDashboardInputState: UserDashboardInputState =
     useUserDashboardInput();
@@ -79,8 +26,11 @@ function App() {
   const {
     lines,
     startDate,
+    setStartDate,
     dayOfWeek,
+    setDayOfWeek,
     endDate,
+    setEndDate,
     updateLinesWithLineMetrics,
     visibleLines,
     isAggregateVisible,
@@ -183,219 +133,53 @@ function App() {
     // eslint-disable-next-line react-hooks/exhaustive-deps
   }, [JSON.stringify(ridershipByLine)]);
 
-  /* Dropping below the breakpoint unmounts DockShell; its api is then dead. */
-  useEffect(() => {
-    if (!isDesktop) {
-      setDockApi(null);
-      setPanelVisibility(allPanelsVisible);
-    }
-  }, [isDesktop]);
-
-  /* Track which panels exist in the dock. */
-  useEffect(() => {
-    if (!dockApi) return;
-
-    const syncVisibility = () => {
-      const present = new Set(dockApi.panels.map((panel) => panel.id));
-      setPanelVisibility(
-        Object.fromEntries(
-          PANEL_IDS.map((id) => [id, present.has(id)]),
-        ) as Record<PanelId, boolean>,
-      );
-    };
-
-    syncVisibility();
-    const addListener = dockApi.onDidAddPanel(syncVisibility);
-    const removeListener = dockApi.onDidRemovePanel(syncVisibility);
-    return () => {
-      addListener.dispose();
-      removeListener.dispose();
-    };
-  }, [dockApi]);
-
-  /* Expand toggle → maximize the line-selector's group. */
-  useEffect(() => {
-    if (!dockApi) return;
-    const panel = dockApi.getPanel('line-selector');
-    if (!panel) return;
-
-    if (isLineSelectorExpanded) {
-      if (!panel.api.isMaximized()) panel.api.maximize();
-    } else if (panel.api.isMaximized()) {
-      panel.api.exitMaximized();
-    }
-  }, [dockApi, isLineSelectorExpanded]);
-
-  /* Maximize changes made through dockview itself → expand toggle state. */
-  useEffect(() => {
-    if (!dockApi) return;
-    const listener = dockApi.onDidMaximizedGroupChange(
-      ({ group, isMaximized }) => {
-        const lineSelectorGroup = dockApi.getPanel('line-selector')?.group;
-        if (group === lineSelectorGroup) {
-          setIsLineSelectorExpanded(isMaximized);
-        } else if (isMaximized) {
-          /* Some other group was maximized; the selector is no longer expanded */
-          setIsLineSelectorExpanded(false);
-        }
-      },
-    );
-    return () => listener.dispose();
-  }, [dockApi]);
-
-  const togglePanel = useCallback(
-    (id: PanelId) => {
-      if (!dockApi) return;
-
-      const existing = dockApi.getPanel(id);
-      if (existing) {
-        dockApi.removePanel(existing);
-        return;
-      }
-
-      const def = PANEL_DEFS[id];
-      let position:
-        | { referencePanel: PanelId; direction: 'above' | 'below' | 'right' }
-        | undefined = def.position;
-      if (position && !dockApi.getPanel(position.referencePanel)) {
-        position = undefined;
-      }
-      if (!position) {
-        /* Reference panel is hidden (or the panel has none): fall back to the
-           first panel still present — above it for the date-range strip,
-           below it otherwise. */
-        const fallbackRef = PANEL_IDS.find(
-          (panelId) => panelId !== id && dockApi.getPanel(panelId),
-        );
-        if (fallbackRef) {
-          position = {
-            referencePanel: fallbackRef,
-            direction: id === 'date-range' ? 'above' : 'below',
-          };
-        }
-      }
-
-      dockApi.addPanel({
-        id,
-        component: def.component,
-        title: def.title,
-        ...(position ? { position } : {}),
-      });
-    },
-    [dockApi],
-  );
-
-  const resetLayout = useCallback(() => {
-    if (!dockApi) return;
-    dockApi.clear();
-    clearLayout();
-    buildDefaultLayout(dockApi);
-    setIsLineSelectorExpanded(false);
-  }, [dockApi]);
-
-  const dockLayoutValue: DockLayoutContextValue = useMemo(
-    () => ({
-      visibility: panelVisibility,
-      togglePanel,
-      resetLayout,
-    }),
-    [panelVisibility, togglePanel, resetLayout],
-  );
-
-  const dashboardValue: DashboardContextValue = {
-    userDashboardInputState,
-    lines,
-    visibleLines,
-    ridershipByLine,
-    chartDatasets,
-    monthList,
-    isLineSelectorExpanded,
-    setIsLineSelectorExpanded,
-  };
-
-  const handleApiReady = useCallback((api: DockviewApi) => {
-    setDockApi(api);
-  }, []);
-
   return (
     /* Stretch full height */
     <div className="flex flex-col min-h-screen mx-4">
       <Header />
 
-      <DashboardProvider value={dashboardValue}>
-        <DockLayoutProvider value={dockLayoutValue}>
-          {isDesktop ? (
-            /* Dockview's root is `height: 100%`, which only resolves against a
-               parent with a definite height — a flex-grown `height: auto` box
-               collapses it to 0. The outer div takes the space; the absolutely
-               positioned inner div gives that space a definite height. */
-            <div className="relative grow min-h-[42rem]">
-              <div className="absolute inset-0">
-                <DockShell
-                  panels={{
-                    'date-range': (
-                      <PanelChrome>
-                        <DateRangePanel />
-                      </PanelChrome>
-                    ),
-                    'line-selector': (
-                      <PanelChrome scroll={false}>
-                        <LineSelectorPanel />
-                      </PanelChrome>
-                    ),
-                    chart: (
-                      <PanelChrome scroll={false}>
-                        <ChartPanel />
-                      </PanelChrome>
-                    ),
-                    summary: (
-                      <PanelChrome>
-                        <SummaryPanel />
-                      </PanelChrome>
-                    ),
-                    map: (
-                      <PanelChrome scroll={false}>
-                        <MapPanel />
-                      </PanelChrome>
-                    ),
-                  }}
-                  onApiReady={handleApiReady}
-                />
-              </div>
-            </div>
-          ) : (
-            /* Below lg: the pre-dock stacked layout, from the same panel content */
-            <>
-              {/* Date range pane */}
-              <div className="pane mb-4">
-                <DateRangePanel />
-              </div>
+      {/* Date range pane */}
+      <div className="pane mb-4">
+        <DateRangeSelector
+          startDate={startDate}
+          setStartDate={setStartDate}
+          endDate={endDate}
+          setEndDate={setEndDate}
+          dayOfWeek={dayOfWeek}
+          setDayOfWeek={setDayOfWeek}
+        />
+      </div>
 
-              <div className="grow flex flex-col gap-4">
-                {/* Metro lines pane */}
-                <div className="pane flex flex-col gap-4 h-[32rem]">
-                  <LineSelectorPanel />
-                </div>
+      {/* Grow to fill remaining vertical space; only one column if expanded or on mobile */}
+      <div
+        className={`grow grid flex-col gap-4 ${isLineSelectorExpanded ? 'lg:grid-cols-[1fr]' : 'grid-cols-[1fr] lg:grid-cols-[25%_1fr]'}`}
+      >
+        {/* Metro lines pane */}
+        {/* Hack to match sibling height - https://www.reddit.com/r/css/comments/15qu1ml/restrict_childs_height_to_parents_height_which_is/*/}
+        <div
+          className={`pane flex flex-col gap-4 h-[32rem] min-h-full w-0 min-w-full ${isLineSelectorExpanded ? 'lg:h-auto' : 'lg:h-0'}`}
+        >
+          <LineSelector
+            {...userDashboardInputState}
+            lines={visibleLines}
+            ridershipByLine={ridershipByLine}
+            isExpanded={isLineSelectorExpanded}
+            setIsExpanded={setIsLineSelectorExpanded}
+          />
+        </div>
 
-                {/* Only show output if the line selector is not expanded */}
-                {!isLineSelectorExpanded && (
-                  <>
-                    <div className="pane">
-                      <ChartPanel />
-                    </div>
-
-                    {chartDatasets.length > 0 && <SummaryPanel />}
-
-                    <div className="pane">
-                      <MapPanel />
-                    </div>
-                  </>
-                )}
-              </div>
-            </>
-          )}
-        </DockLayoutProvider>
-      </DashboardProvider>
+        {/**
+         * Only show right side if line selector not selected
+         * TODO: Change this from conditional rendering to conditional visibility; that way it doesn't rerender every time
+         */}
+        {!isLineSelectorExpanded && (
+          <OutputArea
+            chartDatasets={chartDatasets}
+            months={monthList}
+            lines={lines}
+          />
+        )}
+      </div>
 
       <Footer />
     </div>
