--- src/dock/panels/MapPanel.tsx
@@ -1,8 +1,0 @@
-import Map from '../../components/Map';
-import { useDashboard } from '../../context/DashboardContext';
-
-export default function MapPanel() {
-  const { lines } = useDashboard();
-
-  return <Map lines={lines} />;
-}
